Renamed middleware from authenticate to authorize
