+++ src/http/middleware/index.js
@@ -1,5 +1,5 @@
 module.exports = {
   getScopesQuery: require("./scopes"),
   getIdParam: require("./id"),
-  authenticate: require("./authenticate"),
+  authorize: require("./authorize"),
 }
